Rename examples/websocketchat/wschatresp.js to examples/websocketchat/wschatsend.js
